- sync to 2.0.12

+++ src/gdkanji.c
@@ -391,7 +391,8 @@ do_convert (unsigned char *to, unsigned char *from, const char *code)
   from_len = strlen ((const char *) from) + 1;
   to_len = BUFSIZ;
 
-  if (iconv (cd, (char **) &from, &from_len, (char **) &to, &to_len) == -1)
+  if ((int) (iconv (cd, (char **) &from, &from_len, (char **) &to, &to_len))
+      == -1)
     {
 #ifdef HAVE_ERRNO_H
       if (errno == EINVAL)
